+++ angular/SIG/src/app/Components/newaccount/newaccount.component.html
@@ -1,32 +1,34 @@
 <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
 
 <style>
-    body {
-    background-color: rgb(235, 238, 243);
-    }
-  </style>
+  body {
+  background-color: rgb(235, 238, 243);
+  }
+</style>
   
-  <div class=rsection style="text-align: center;">
-    <h2>Let's get started. Create new account below.</h2>
-  </div>
+<div class=rsection style="text-align: center;">
+  <h2>Let's get started. Create new account below.</h2>
+</div>
 
-  <form style="text-align: center;">
-      <div class="form-row">
-        <div class="form-group col-md-4">
-          <label for="inputSelectAccountType">Selcect Account Type</label>
-          <select id="inputSelectAccountType" class="form-control">
-            <option selected>Brokerage</option>
-            <option>Roth IRA</option>
-          </select>
-        </div>
-        <div class="form-group col-md-4">
-          <label for="inputSymbol">Initial Funding Amount</label>
-          <input type="text" class="form-control" id="inputSymbol" placeholder="$">
-        </div>
+<form style="text-align: center;">
+    <div class="form-row">
+      <div class="form-group col-md-4">
+        <label for="inputSelectAccountType">Selcect Account Type</label>
+        <select id="inputSelectAccountType" class="form-control">
+          <option selected>Brokerage</option>
+          <option>Roth IRA</option>
+        </select>
       </div>
-      <button type="submit" class="btn btn-primary">Create Account</button>
-      <a href="#" class="btn btn-primary">Create Account</a>
-    </form>
+      <div class="form-group col-md-4">
+        <label for="inputSymbol">Initial Funding Amount</label>
+        <input type="text" class="form-control" id="inputSymbol" placeholder="$">
+      </div>
+    </div>
+<button type="submit" class="btn btn-primary">Create Account</button>
+<a href="#" class="btn btn-primary">Create Account</a>
+</form>
+
+
 
 <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
