+++ angular/SIG/src/app/Components/orders/orders.component.html
@@ -17,36 +17,47 @@
   </div>
 </div>
 
-
 <form style="text-align: center;">
+  <div class="col-md-6 offset-md-3">
+    <div class="card text-center">
+      <div class="card-header">
+        GOOG
+      </div>
+      <div class="card-body">
+        <h5 class="card-title">1,103</h5>
+      </div>
+      <div class="card-footer text-muted">
+        Volume 2.00M
+      </div>
+    </div>
+  </div>
   <div class="form-row">
-    <div class="form-group col-md-4">
-      <label for="inputTransactionType">Transaction Type</label>
-      <select id="inputTransactionType" class="form-control">
-        <option selected>Stocks/ETFs</option>
-        <option>Mutual Funds</option>
-        <option>Bonds</option>
-      </select>
-    </div>
-    <div class="form-group col-md-4">
-      <label for="inputSymbol">Symbol</label>
-      <input type="text" class="form-control" id="inputSymbol">
-    </div>
+      <div class="form-group col-md-4">
+        <label for="inputTransactionType">Action</label>
+        <select id="inputTransactionType" class="form-control">
+          <option selected>Buy</option>
+          <option>Sell</option>
+          <option>Trade</option>
+        </select>
+      </div>
+      <div class="form-group col-md-4">
+        <label for="inputSymbol">Quantity</label>
+        <input type="text" class="form-control" id="inputSymbol">
+      </div>
   </div>
-  <div class="card text-center">
-    <div class="card-header">
-      GOOG
+  <p class="text-center">Estimated Order Value</p>
+  <p class="text-center">Estimated Commission $10.00</p>
+  <div class="col-md-6 offset-md-3">
+    <div class="alert alert-primary" role="alert">
+      <div class="font-weight-bold">
+        Estimated Total Value
+      </div>
     </div>
-    <div class="card-body">
-      <h5 class="card-title">1,103</h5>
-    </div>
-    <div class="card-footer text-muted">
-      Volume 2.00M
-    </div>
-  </div>
+  </div>  
   <button type="submit" class="btn btn-primary">Place Order</button>
   <a href="#" class="btn btn-primary">Place Order</a>
 </form>
+
 
 
 <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
